requirements.txt
streamlit
opencv-python-headless
numpy
pandas
scikit-learn
scikit-image
pillow
